requirements.txt
streamlit>=1.21.0
altair<5
pytest
anthropic>=0.25.0
python-dotenv>=1.0.0
